refactored cli commands into separate files
